% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agecalc.R
\name{agecalc}
\alias{agecalc}
\title{Calculate the age of the target day}
\usage{
agecalc(birthday, target_day = today())
}
\arguments{
\item{birthday}{For example "1965/09/13" or "1965.09.13" or "1965-09-13"}

\item{target_day}{Date for calculating age. If you omit the argument, today is entered.}
}
\description{
Calculate the age of the target day
}
