% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmicalc.R
\name{bmicalc}
\alias{bmicalc}
\title{Calculate body mass index}
\usage{
bmicalc(height, weight)
}
\arguments{
\item{height}{Unit is centimeter}

\item{weight}{Unit is kilogram}
}
\description{
Calculate body mass index
}
