% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsacalc.R
\name{bsacalc}
\alias{bsacalc}
\title{Calculate body surface area}
\usage{
bsacalc(height, weight, formula = 1)
}
\arguments{
\item{height}{Unit is centimeter}

\item{weight}{Unit is kilogram}

\item{formula}{1:DuBois 2:Shintani 3:Fujimoto}
}
\description{
Calculate body surface area
}
