% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrcalc.R
\name{ccrcalc}
\alias{ccrcalc}
\title{Calculate creatinine clearance}
\usage{
ccrcalc(age, weight, scr, gender, correction = 1)
}
\arguments{
\item{age}{Age of the subject}

\item{weight}{Unit is kilogram}

\item{scr}{Serum creatinine value. Unit is mg/dL.}

\item{gender}{female or male}

\item{correction}{When the value is 1, if scr is less than 0.6, scr is rounded up to 0.6. If the value is 0, it will not be rounded up. The default value is 1.}
}
\description{
Calculate creatinine clearance
}
