% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdwtcalc.R
\name{stdwtcalc}
\alias{stdwtcalc}
\title{Calculate standard weight}
\usage{
stdwtcalc(height, weight)
}
\arguments{
\item{height}{Unit is centimeter}

\item{weight}{Unit is kilogram}
}
\description{
Calculate standard weight
}
