% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wareki2seireki.R
\name{wareki2seireki}
\alias{wareki2seireki}
\title{Convert Japanese calendar to Christian calendar year}
\usage{
wareki2seireki(wareki)
}
\arguments{
\item{wareki}{The value of Japanese calendar. For example, please input like "S47.03.31".}
}
\description{
This function converts the Japanese calendar to the Christian era. It corresponds to Meiji, Taisho, Showa, Heisei. For example, converts "S47.03.31" to "1972/03/31".
}
