% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agecalc.R
\name{agecalc}
\alias{agecalc}
\title{Calculate the age of the target day}
\usage{
agecalc(target_day, birthday)
}
\arguments{
\item{target_day}{Date for calculating age}

\item{birthday}{For example "1965/09/13" or "1965.09.13" or "1965-09-13"}
}
\description{
Calculate the age of the target day
}
